src/app/employee-form/employee-form.component.ts
import { Component } from '@angular/core';
import { EmployeeService } from '../services/employee.service';
import { Employee } from '../models/employee';
import { FormsModule } from '@angular/forms';

@Component({
  selector: 'app-employee-form',
  standalone: true,
  imports: [FormsModule],
  templateUrl: './employee-form.component.html',
  styleUrl: './employee-form.component.css'
})
export class EmployeeFormComponent {
  newEmployee: Employee = { id: 0, name: '', role: '' };

  constructor(private employeeService: EmployeeService) {}

  addEmployee(): void {
    this.employeeService.addEmployee(this.newEmployee).subscribe(() => {
      alert('Employee added successfully');
      window.location.reload(); // Refresh the list
    });
  }

}
